new_structure/requirements.txt
python-dotenv==0.19.2
numpy>=1.26.0
pandas>=2.0.0
pyupbit==0.2.33
anthropic==0.3.6
schedule==1.1.0
psutil==5.8.0
selenium==3.141.0
streamlit==0.88.0
matplotlib